fix most warnings, started tidying up files
